Move login to /auth/login, securize /posts routes
